Implement a new ./pant vex task for distributing Python projects.
The vex task takes a single PythonBinary as input and produces a .tar.gz of a "Vex" as output. This is then distributed to target computers, and the proper "Vex" directory is extracted. A Vex is a "Virtual Environment eXecutable". Once extracted, a Vex directory structure will have two important scripts in it. A user at a target computer first runs the build-vex.sh script to bootstrap a virtual environment and install the PythonBinary into it. Then that user executes run.sh to execute the PythonBinary's source code.